// src/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IProjectFields {
	/** Title */
	title: string;

	/** Slug */
	slug: string;

	/** Short description */
	shortDescription: string;

	/** Month */
	month:
		| "Januar"
		| "Februar"
		| "Mars"
		| "April"
		| "Mai"
		| "Juni"
		| "Juli"
		| "August"
		| "September"
		| "Oktober"
		| "November"
		| "Desember";

	/** Year */
	year: number;

	/** Categories */
	categories: ("Design" | "Kode")[];

	/** Blurb */
	blurb: string;

	/** Link */
	link?: string | undefined;

	/** GitHub-link */
	gitHubLink?: string | undefined;

	/** Thumbnail */
	thumbnail: Asset;

	/** Images */
	images?: Asset[] | undefined;
}

export interface IProject extends Entry<IProjectFields> {
	sys: {
		id: string;
		type: string;
		createdAt: string;
		updatedAt: string;
		locale: string;
		contentType: {
			sys: {
				id: "project";
				linkType: "ContentType";
				type: "Link";
			};
		};
	};
}

export type CONTENT_TYPE = "project";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";